Fixed some warnings; Fixed an issue in CMake where USE_STATIC_RUNTIME was not properly applied when the engine was the top CMake project; Ensured the FEBasicApplication CMake project is included before checking USE_STATIC_RUNTIME.

+++ ResourceManager/FEResourceManager.cpp
@@ -803,7 +803,7 @@ bool FEResourceManager::ExportFEMeshToOBJ(FEMesh* MeshToExport, const char* File
 		// For each triangle's vertex
 		for (size_t j = 0; j < 3; j++)
 		{
-			size_t CurrentIndex = Indices[i + j] + 1; // +1 because OBJ indices start at 1
+			int CurrentIndex = Indices[i + j] + 1; // +1 because OBJ indices start at 1
 
 			// In OBJ, each face vertex is defined as v/vt/vn
 			Data.RawIndices.push_back(CurrentIndex); // vertex position index
